Symbian compile fix
armv5 complained that the call to QChar::category() was ambiguous,
because there is a version taking an uint, and a version taking ushort as
argument.

+++ src/script/bridge/qscriptdeclarativeclass.cpp
@@ -473,7 +473,7 @@ bool QScriptDeclarativeClass::startsWithUpper(const Identifier &identifier)
     JSC::UString::Rep *r = (JSC::UString::Rep *)identifier;
     if (r->size() < 1)
         return false;
-    return QChar::category(r->data()[0]) == QChar::Letter_Uppercase;
+    return QChar::category((ushort)(r->data()[0])) == QChar::Letter_Uppercase;
 }
 
 quint32 QScriptDeclarativeClass::toArrayIndex(const Identifier &identifier, bool *ok)
